refactor: implement chrome.i18n.getMessage() without the remote module
