+++ lib/browser/chrome-extension.js
@@ -4,9 +4,11 @@ const { app, webContents, BrowserWindow } = require('electron')
 const { getAllWebContents } = process.atomBinding('web_contents')
 const renderProcessPreferences = process.atomBinding('render_process_preferences').forAllWebContents()
 const { ipcMainInternal } = require('@electron/internal/browser/ipc-main-internal')
+const ipcMainUtils = require('@electron/internal/browser/ipc-main-internal-utils')
 
 const { Buffer } = require('buffer')
 const fs = require('fs')
+const os = require('os')
 const path = require('path')
 const url = require('url')
 
@@ -166,10 +168,6 @@ ipcMainInternal.on('CHROME_RUNTIME_CONNECT', function (event, extensionId, conne
   page.webContents._sendInternalToAll(`CHROME_RUNTIME_ONCONNECT_${extensionId}`, event.sender.id, portId, connectInfo)
 })
 
-ipcMainInternal.on('CHROME_I18N_MANIFEST', function (event, extensionId) {
-  event.returnValue = manifestMap[extensionId]
-})
-
 let resultID = 1
 ipcMainInternal.on('CHROME_RUNTIME_SENDMESSAGE', function (event, extensionId, message, originResultID) {
   const page = backgroundPages[extensionId]
@@ -199,6 +197,34 @@ ipcMainInternal.on('CHROME_TABS_SEND_MESSAGE', function (event, tabId, extension
     event._replyInternal(`CHROME_TABS_SEND_MESSAGE_RESULT_${originResultID}`, result)
   })
   resultID++
+})
+
+const getMessagesPath = (extensionId, language) => {
+  const metadata = manifestMap[extensionId]
+  if (!metadata) {
+    throw new Error(`Invalid extensionId: ${extensionId}`)
+  }
+
+  const isSafe = path.dirname(path.resolve(os.tmpdir(), language)) === os.tmpdir()
+  if (!isSafe) {
+    throw new Error(`Invalid language: ${language}`)
+  }
+
+  const localesDirectory = path.join(metadata.srcDirectory, '_locales')
+
+  try {
+    const filename = path.join(localesDirectory, language, 'messages.json')
+    fs.accessSync(filename, fs.constants.R_OK)
+    return filename
+  } catch (err) {
+    const defaultLocale = metadata.default_locale || 'en'
+    return path.join(localesDirectory, defaultLocale, 'messages.json')
+  }
+}
+
+ipcMainUtils.handleSync('CHROME_GET_MESSAGES', function (event, extensionId, language) {
+  const messagesPath = getMessagesPath(extensionId, language)
+  return fs.readFileSync(messagesPath)
 })
 
 const isChromeExtension = function (pageURL) {
