+++ lib/renderer/extensions/i18n.js
@@ -6,37 +6,12 @@
 // Does not implement predefined messages:
 // https://developer.chrome.com/extensions/i18n#overview-predefined
 
-const { potentiallyRemoteRequire } = require('@electron/internal/renderer/remote')
-const ipcRenderer = require('@electron/internal/renderer/ipc-renderer-internal')
-const fs = potentiallyRemoteRequire('fs')
-const path = potentiallyRemoteRequire('path')
-
-let metadata
-
-const getExtensionMetadata = (extensionId) => {
-  if (!metadata) {
-    metadata = ipcRenderer.sendSync('CHROME_I18N_MANIFEST', extensionId)
-  }
-  return metadata
-}
-
-const getMessagesPath = (extensionId, language) => {
-  const metadata = getExtensionMetadata(extensionId)
-  const localesDirectory = path.join(metadata.srcDirectory, '_locales')
-  try {
-    const filename = path.join(localesDirectory, language, 'messages.json')
-    fs.accessSync(filename, fs.constants.R_OK)
-    return filename
-  } catch (err) {
-    const defaultLocale = metadata.default_locale || 'en'
-    return path.join(localesDirectory, defaultLocale, 'messages.json')
-  }
-}
+const ipcRendererUtils = require('@electron/internal/renderer/ipc-renderer-internal-utils')
 
 const getMessages = (extensionId, language) => {
   try {
-    const messagesPath = getMessagesPath(extensionId, language)
-    return JSON.parse(fs.readFileSync(messagesPath)) || {}
+    const data = ipcRendererUtils.invokeSync('CHROME_GET_MESSAGES', extensionId, language)
+    return JSON.parse(data) || {}
   } catch (error) {
     return {}
   }
